#10 Publish new clinics' availability after a booking occupied the last available slot for a clinic
When a successful booking is made and in the event that this reserved the last available time slot for that date for that specific dentist clinic, then the new data is published in order to get updated on the client's map markers afterwards.
Refactor filesystem readFile method calls from asynchronous processing to synchronous one in timeSlotController.js and changed the process in storageController.

+++ src/availability-handler/storageController.js
@@ -24,10 +24,11 @@ class StorageController {
     saveAvailability(availability, id) {
         let fileName = './availability-data/availability-' + id +'.json'
 
-        fs.readFile(fileName, () => {
+        if (fs.existsSync(fileName)) {
             fs.writeFileSync(fileName, JSON.stringify(availability));
-            console.log('booking saved ')
-        })
+        } else {
+            console.log('Error: the file that was specified in order to save the available time slots does not exist.')
+        }
     }
 }
 module.exports.StorageController = StorageController
